remove spec_excludes from test_spec_exclude_integration
